# Update benchmark sciprt for benchmark_driver 0.9.0