--- a/string_template.gemspec
+++ b/string_template.gemspec
@@ -24,5 +24,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'bundler'
   spec.add_development_dependency 'rake'
   spec.add_development_dependency 'minitest'
-  spec.add_development_dependency 'benchmark_driver'
+  spec.add_development_dependency 'benchmark_driver', '>= 0.9.0'
 end